==> 04-missing.Rmd <==
# Missing values

## Dataset 1 Obesity-Behavioral

This dataset contained two broad categories of missing values - some for which the rows were completely missing from the data, and others which had supplementary text in particular columns explaining that there was missing data for the given row. 

``` {r}
library(dplyr)
library(ggplot2)
library(tidyr)

brfs <- read.csv("data/clean/BRFS_Filtered.csv")

```

```{r}

library(extracat)

brfs2 <- brfs
brfs2[brfs2 == "Data not reported" ] <- NA

# Missing values
visna(brfs2)

```

In above plot, we observed two patterns of missing data. 

One is where due to "insufficient sample size" (mentioned in the Data_Value_Footnote column) for a given state and Question, some data values are missing.

We also noted that within the "Income" StratificationCategory1 group, in the "Stratification1" and "Income" columns, a subcategory called "Data not reported" constituted a small proportion of the entire dataset. In our further analysis, we treat this strata as just another subcategory within "Income".

Apart from these obvious missing patterns, this dataset also has completely missing rows of data across different years. Since these rows are not a part of the dataset explicitly, the above plot failed to capture them. We added rows for these missing values and indicated the Data_Value as missing, to enable us to visualise patterns of these missing values. This is shown in below plot.

```{r, fig.height=7, fig.width=10}

# convert to % of country and class total
brfs_missing <- brfs %>%
                group_by(YearStart,QuestionRenamed ) %>% 
                summarise(Total = n()) %>%
                pivot_wider(names_from = QuestionRenamed, values_from = Total) %>%
                pivot_longer(!YearStart, names_to = "QuestionRenamed", values_to = "Total") %>%
                mutate(Missing=is.na(Total))

# make custom theme
theme_heat <- theme_classic() +
  theme(axis.line = element_blank(),
        axis.ticks = element_blank())

# basic plot
gtile <- ggplot(brfs_missing, aes(x = YearStart, y = QuestionRenamed, fill=Missing)) +
  geom_tile(color="white") +
  scale_x_continuous(breaks = seq(2011 , 2020 , by=1)) +
  scale_fill_viridis_d() +
  theme_heat +
  ggtitle("Missing Values in Obesity Behavioural Data for all Classes") +
  xlab("Year") + 
  ylab("Topic")

plot(gtile)

```
Data for Nutrition Status i.e. Questions - "No Vegetable Consumption" and "No Fruit Consumption" is missing from 2011-2015 and 2016.

Data for four of the Physicial Activity Status Questions is also missing for alternate years starting from 2012 to 2018. 

Note that above heatmap only captures the missing entries that we explicitly added back - and not the NA values that the previous plot displayed.

## Dataset 2 Obesity-Policy-Support

```{r, fig.height=7, fig.width=10}

pe <- read.csv("data/clean/PE_Filtered.csv")

# convert to % of country and class total
pe_missing <- pe %>%
                group_by(YearStart,QuestionRenamed ) %>% 
                summarise(Total = n()) %>%
                pivot_wider(names_from = QuestionRenamed, values_from = Total) %>%
                pivot_longer(!YearStart, names_to = "QuestionRenamed", values_to = "Total") %>%
                mutate(Missing=is.na(Total))

# basic plot
gtile2 <- ggplot(pe_missing, aes(x = YearStart, y = QuestionRenamed, fill=Missing)) +
  geom_tile(color="white") +
  scale_x_continuous(breaks = seq(2011 , 2020 , by=1)) +
  scale_fill_viridis_d() +
  theme_heat +
  ggtitle("Missing Values in Obesity Policy Support Data for all Classes") +
  xlab("Year") + 
  ylab("Topic")

plot(gtile2)

```
For the Obesity-Policy-Support Dataset, we did a similar analysis.
There are no NA values in this dataset.

To capture the missing entries over the years, we employed the same technique, as used for the Obesity-Behavioural dataset to generate this heatmap of missing entries.

No distinct patterns can be found. However, overall, this dataset seems much more sparse than the Obesity-Behavioural dataset.

As compared to 2019, 2018 has lesser number of missing values. Hence, in all our analysis at the state-level for the policy related questions, we used data from 2018 (instead of 2019 which we used from Behavioral Obesity Dataset1) as recent data.
