==> 05-results.Rmd <==
# Results

``` {r}
library(dplyr)
library(ggplot2)
library(tidyr)
library(forcats)
library(directlabels)
library(choroplethr)
library(choroplethrMaps)
library(statebins)
library(viridis)

brfs <- read.csv("data/clean/BRFS_Filtered.csv")
pe <- read.csv("data/clean/PE_Filtered.csv")

```

## Dataset 1 Obesity-Behavioral

In this dataset, we have data from 2011-2019. This data was utilised to form a time-series graph showing the trends of Obesity Rates over the last decade, in different StratificationCategories and also SubCategories.

``` {r, fig.height=10, fig.width=10}

brfs_national <- brfs %>% filter(LocationDesc=="National")

brfs_national_obesity <- brfs_national %>% filter(Class == "Obesity / Weight Status" & QuestionRenamed=="Obese")

brfs_national_obesity_filtered <-brfs_national_obesity %>% select(YearStart,StratificationCategory1,Stratification1,Data_Value)

brfs_national_obesity_filtered$StratificationCategory1 <- as.factor(brfs_national_obesity_filtered$StratificationCategory1)

brfs_national_obesity_filtered$Stratification1 <- as.factor(brfs_national_obesity_filtered$Stratification1)

brfs_national_obesity_filtered$Stratification1 <- relevel(brfs_national_obesity_filtered$Stratification1,"Less than $15,000")

brfs_national_obesity_filtered$Stratification1 <- fct_relevel(brfs_national_obesity_filtered$Stratification1, "Less than high school", after = 1)

brfs_national_obesity_filtered$Stratification1 <- fct_relevel(brfs_national_obesity_filtered$Stratification1, "College graduate", after = Inf)

brfs_national_obesity_filtered <- brfs_national_obesity_filtered %>% rename(Stratification=Stratification1)

# make list of plots
ggList <- lapply(split(brfs_national_obesity_filtered, brfs_national_obesity_filtered$StratificationCategory1), function(i) {
  ggplot(i, aes(y=Data_Value, x=YearStart))+geom_point(aes(colour = Stratification, group = Stratification))+geom_line(aes(colour = Stratification, group = Stratification))+xlab("Year")+ylab("Data_Value in %")+ scale_x_continuous(breaks = seq(2011,2020,2))}) 

# plot as grid in 1 columns
cowplot::plot_grid(plotlist = ggList, ncol = 2,
                   align = 'v', labels = levels(brfs_national_obesity_filtered$StratificationCategory1))

```

For each of the Stratification Categories, we can observe a clear upwards trend i.e. as the years have gone by, the obesity rates have increased. The "Total" plot summarises this trend across all Stratification Categories. 

With the "Age(years)", "Education", "Income" and "Race/Ethnicity" Categories, we can see a visible difference in obesity rates of some sub-groups within them (lines located at the bottom part of the graph, ex: College graduates have much lower obesity rates than less educated adults), over the others. In our analysis below, we continue to look deeper into these groups.


``` {r, fig.height=7, fig.width=10}

questions <- c("Obese","No Activity","No Fruit Consumption", "No Vegetable Consumption")

# Filter year and questions first
brfs_2019_filtered <-brfs_national %>%
  filter(YearStart==2019 & QuestionRenamed %in% questions) %>%
  select(YearStart,LocationDesc,Class,StratificationCategory1,Stratification1,QuestionRenamed,Data_Value)

brfs_2019_filtered$Class2 <- as.factor(brfs_2019_filtered$Class)
brfs_2019_filtered$Class2 <- relevel(brfs_2019_filtered$Class2,"Obesity / Weight Status")
levels(brfs_2019_filtered$Class2) <- c("Obesity Status","No Fruit/Vegetable Consumption Daily","No Physical Activity Weekly")

cols <- c("Class Average %"="red")

brfs_2019_age <-brfs_2019_filtered %>%
  filter(StratificationCategory1=="Age (years)") %>%
  group_by(Class2,Stratification1) %>%
  summarise(mean_data_value = mean(Data_Value))

brfs_2019_age.mean = brfs_2019_age %>%
  group_by(Class2) %>%
  mutate(ymean = mean(mean_data_value))

barplot1 <-  ggplot(brfs_2019_age,aes(x=Stratification1, y=mean_data_value))+
  geom_col(fill="lightBlue")+geom_text(aes(label=mean_data_value),vjust = -1.5, position = position_dodge(width = 1),size=3.5)+
  geom_errorbar(data=brfs_2019_age.mean, aes(x=Stratification1, ymax = ymean, ymin = ymean, colour="Class Average %"),
               size=0.5, linetype = "longdash", inherit.aes = F, width = 1)+
  scale_colour_manual(name="Dotted Line",values=cols)+ scale_fill_manual(name="Bar",values=cols) +
  facet_wrap(~Class2, ncol = 1) + ylim(c(0,100))+
  ylab("Percentage") +
  xlab("Age (years) ") +
  ggtitle("Percentage across different Age Groups (years) in 2019") 

plot(barplot1)

```

We can see that middle bars are highest. For lower age group obesity is least - although fruit/vegetable consumption is low, physical activity is higher. As age increases, as expected, the physical activity is the least, although overall nutrition and obesity status are better than middle groups.

``` {r, fig.height=7, fig.width=10}

brfs_2019_income <-brfs_2019_filtered %>%
  filter(StratificationCategory1=="Income") %>%
  group_by(Class2,Stratification1) %>%
  summarise(mean_data_value = mean(Data_Value))

brfs_2019_income$Stratification1 <- fct_relevel(brfs_2019_income$Stratification1,"Less than $15,000")

brfs_2019_income.mean = brfs_2019_income %>%
  group_by(Class2) %>%
  mutate(ymean = mean(mean_data_value))

barplot2 <-  ggplot(brfs_2019_income,aes(x=Stratification1, y=mean_data_value))+
  geom_col(fill="lightBlue")+geom_text(aes(label=mean_data_value),vjust = -1.5, position = position_dodge(width = 1),size=3.5)+
  geom_errorbar(data=brfs_2019_income.mean, aes(x=Stratification1, ymax = ymean, ymin = ymean, colour="Class Average %"),
               size=0.5, linetype = "longdash", inherit.aes = F, width = 1)+
  scale_colour_manual(name="Dotted Line",values=cols)+ scale_fill_manual(name="Bar",values=cols) +
  facet_wrap(~Class2, ncol = 1) + ylim(c(0,100))+
  ylab("Percentage") +
  xlab("Income Groups") +
  ggtitle("Percentage across different Income Groups in 2019")

plot(barplot2)

```

We observe an overall decreasing trend. Population belonging to higher income levels are found to be more physically active, eat more nutritious food, and also have better obesity status than lower income groups.

``` {r, fig.height=7, fig.width=10}

brfs_2019_education <-brfs_2019_filtered %>%
  filter(StratificationCategory1=="Education") %>%
  group_by(Class2,Stratification1) %>%
  summarise(mean_data_value = mean(Data_Value))

brfs_2019_education.mean = brfs_2019_education %>%
  group_by(Class2) %>%
  mutate(ymean = mean(mean_data_value))

brfs_2019_education$Stratification1 <- fct_relevel(brfs_2019_education$Stratification1, "Less than high school", after = 1)

brfs_2019_education$Stratification1 <- fct_relevel(brfs_2019_education$Stratification1, "College graduate", after = Inf)

barplot3 <-  ggplot(brfs_2019_education,aes(x=Stratification1, y=mean_data_value))+
  geom_col(fill="lightBlue")+geom_text(aes(label=mean_data_value),vjust = -1.5, position = position_dodge(width = 1),size=3.5)+
  geom_errorbar(data=brfs_2019_education.mean, aes(x=Stratification1, ymax = ymean, ymin = ymean, colour="Class Average %"),
               size=0.5, linetype = "longdash", inherit.aes = F, width = 1)+
  scale_colour_manual(name="Dotted Line",values=cols)+ scale_fill_manual(name="Bar",values=cols) +
  facet_wrap(~Class2, ncol = 1) + ylim(c(0,100))+
  ylab("Percentage") +
  xlab("Education Levels") + 
  ggtitle("Percentage across different Education Groups in 2019")

plot(barplot3)

```

An overall decreasing trend is apparent. As education level increases, there is more physical activity taking place, more nutritious food being consumed, and more people with higher educationals levels are seen to have a better obesity status.

``` {r, fig.height=7, fig.width=10}

# Filter for race data

brfs_2019_race <-brfs_2019_filtered %>%
  filter(StratificationCategory1=="Race/Ethnicity") %>%
  group_by(Class2,Stratification1) %>%
  summarise(mean_data_value = mean(Data_Value))

brfs_2019_race.mean = brfs_2019_race %>%
  group_by(Class2) %>%
  mutate(ymean = mean(mean_data_value))

cols <- c("Class Average %"="red")

barplot4 <-  ggplot(brfs_2019_race,aes(x=Stratification1, y=mean_data_value))+
  geom_col(fill="lightBlue")+geom_text(aes(label=mean_data_value),vjust = -1.5, position = position_dodge(width = 1),size=3.5)+
  geom_errorbar(data=brfs_2019_race.mean, aes(x=Stratification1, ymax = ymean, ymin = ymean, colour="Class Average %"),
               size=0.5, linetype = "longdash", inherit.aes = F, width = 1)+
  scale_colour_manual(name="Dotted Line",values=cols)+ scale_fill_manual(name="Bar",values=cols) +
  facet_wrap(~Class2, ncol = 1) + ylim(c(0,100))+
  ylab("Percentage") +
  xlab("Race/Ethnicity Groups") + 
  ggtitle("Percentage across different Race/Ethnicity Groups in 2019") +
  theme(axis.text.x = element_text(angle=45, hjust = 1))

plot(barplot4)
```

For Asians, the obesity status is the least. The levels of "Lack of Physical Activity" and "Lack of Consumption of Fruits and Vegetables" are correspondingly lesser than other races. However, we also note that these numbers are not that much higher than Asians for Hawaiian/Pacific Island people even though their obesity status is the highest. This could be due to genetic or other factors, not captured in the available datasets.
For the sub-categories of 2 or more races, American Indian/Alaska Native, Non-Hispanic Black, we see that they have less nutrition and physical activity than others, and also high obesity status.

Until now, we were looking at Total data across the US states. Now let's look at the data at the state-level and see if we spot any distinctive patterns.

``` {r, fig.height=7, fig.width=10}

brfs_2019_state <- brfs %>%
    filter(YearStart==2019 & 
           Question=="Percent of adults aged 18 years and older who have obesity" &
           StratificationCategory1=="Total")

df_brfs <- brfs_2019_state %>% as.data.frame() %>% 
    transmute(region = tolower(LocationDesc), value = Data_Value)

state_choropleth(df_brfs,
                 title = "State Obesity Percent, 2019",
                 legend = "Percent Obese")

```

We can clearly see that there exists a chain of Southern states, starting from Oklahoma to West Virginia, with the highest Percentage of Obese Adults among all states. Now, let us try to dig deeper into the data for these states - let us look at how they fare in terms of Physical Activity and Nutrition, and if we can find an association with their corresponding Obesity Rates.

```{r,fig.height=7, fig.width=10}

brfs_2019_state_all <-brfs %>%
  filter(YearStart==2019 & QuestionRenamed %in% questions & StratificationCategory1=="Total") %>%
  select(YearStart,LocationDesc,Class,StratificationCategory1,Stratification1,QuestionRenamed,Data_Value)

brfs_2019_state_all <- brfs_2019_state_all %>% 
                        group_by(LocationDesc, Class) %>% 
                            summarise(mean_data_value = mean(Data_Value))

ggplot(brfs_2019_state_all, aes(mean_data_value, fct_reorder2(LocationDesc, Class=="Obesity / Weight Status", mean_data_value, .desc=FALSE), color=Class)) +
  geom_point() +
  #scale_color_manual(values = c("blue", "red")) +
  xlab("Percent of People") +
    ylab("State") +
    ggtitle("Percent of Obesity And Its Factors Per State")

```
The first thing we can notice that the overall trend for "Lack of Physical Activity" and "Lack of Consumption of Fruits and Vegetables" mirrors the overall trend of Obesity / Weight Status, especially for the upper half the graph which represents states with the Highest obesity percentages in the US.

Combining our findings from the previous geo-plot, we find that the Southern States that showed up as having highest obesity rates also seem to be doing worse than other states in terms of Physical Activity and Nutrition - their rates of the Lack of Physical Activty and the Lack of Fruit/Vegetable Consumption are higher than other states.

From this dataset, our broad conclusions are:

1. Overall increasing trend of Obesity regardless of Stratification Category, over the past decade

2. Some sub-groups within each Stratification Category are doing better with their Obesity Status than others. In most cases, this is correlated with their lower rates of Lack of Physical Activty and Nutrition Intake.

3. Southern states in the US have the highest Obesity Rates. This seems to be closely associated with their higher rates of Lack of Physical Activity and Nutrition Intake.

## Dataset 2 Obesity-Policy-Support 

CDC lists proven strategies that support healthy eating and physcial activty, breastfeeding in child care etc. [1] This dataset helps us understand the Support that the US government provides across different states. First, we try to visualise the compliance of the State with the National policies for the given classes of proven strategies.

``` {r, fig.height=6, fig.width=6}

pe_2018_states <- pe %>% filter(YearStart==2018 & !LocationAbbr=="US")

# Nutrition Histogram
questions_nutrition_yesnoq <- c("State-Level Food Policy Council","State-National Alignment on Fruit Serving  Regulations","State-National Alignment on Vegetable Serving Regulations","State-Level Farm to School/Preschool Policy")

pe_2018_nutrition_yesnoq <- pe_2018_states %>%
  filter(Class=="Fruits and Vegetables" & QuestionRenamed %in% questions_nutrition_yesnoq) %>%
  group_by(LocationAbbr) %>%
  summarise(n_policies = sum(Data_Value=="Yes"))

hist <- ggplot(pe_2018_nutrition_yesnoq, aes(x= n_policies)) +
  geom_histogram(binwidth=1,center=1, colour = "black", fill="#f1beb6", closed="left") +
  xlab("No. of Policies for Better Nutrition at the State-Level") +
  ylab("No. of States")+
  ggtitle("No. of Policies for Better Nutrition at the State-Level, 2018") +
  scale_x_continuous(breaks = seq(0 , 5 , by=1)) +
  ylim(c(0,40)) +
  stat_bin(binwidth=1,center=1,geom="text", aes(label=..count..), vjust=-1.5, size=3.5)+
  theme_minimal()

plot(hist)

```

Majority of the states (30) containing exactly 1 Policy Encouraging Better Nutrition in that state. Only 5 out of 41 states contained >=2 policies supporting better nutrition.

``` {r, fig.height=6, fig.width=6}

# Distribution of counts of Local Food Policy Councils across States
pe_2018_nutrition_localpolicycouncil_states <- pe_2018_states %>%
  filter(Class=="Fruits and Vegetables" & QuestionRenamed =="State-Level Food Policy Council"
         & Data_Value=="Yes") %>%
  select(LocationAbbr) %>%
  distinct()

pe_2018_nutrition_localpolicycouncil_counts <- pe_2018_states %>%
  filter(LocationAbbr %in% pe_2018_nutrition_localpolicycouncil_states$LocationAbbr & QuestionRenamed=="Local Food Policy Council Count") %>%
  select(LocationAbbr,Data_Value) %>%
  distinct()

pe_2018_nutrition_localpolicycouncil_counts$Data_Value <- as.numeric(pe_2018_nutrition_localpolicycouncil_counts$Data_Value)

hist_localpolicycounts <- ggplot(pe_2018_nutrition_localpolicycouncil_counts, aes(x= Data_Value))+
  geom_histogram(binwidth=5,center=2.5, colour = "black", fill="#f1beb6", closed="left") +
  xlab("No. of Food Policy Councils")+
  ylab("No. of States")+
  ggtitle("Distribution of No. of Local Food Policy Counts per State, 2018") +
  scale_x_continuous(breaks = seq(0 , 32 , by=5)) +
  stat_bin(binwidth=5,center=2.5,geom="text", aes(label=..count..), vjust=-1.5, size=3.5, closed="left")+
  theme_minimal()

plot(hist_localpolicycounts)

```

In the States that do have Food Policy Councils, most of them seem to have quite a low number - four or less.

``` {r, fig.height=6, fig.width=6}

# Physical Activity
questions_pa_yesnoq <- c("State-National Alignment on Physical Activity for Preschoolers","Complete Streets Policy",
                         "State Guidance on Policies for School Activity Facilities","State Guidance on Policies for School Recess","State Guidance on Policies for Time Spent in Activity During PE Class","State Guidance on Commuting to School by Walk or Bike")
pe_2018_pa_yesnoq <- pe_2018_states %>%
  filter(Class=="Physical Activity" & QuestionRenamed %in% questions_pa_yesnoq) %>%
  group_by(LocationAbbr) %>%
  summarise(n_policies = sum(Data_Value=="Yes"))

hist2 <- ggplot(pe_2018_pa_yesnoq, aes(x= n_policies)) +
  geom_histogram(binwidth=1,center=1, colour = "black", fill="#f1beb6", closed="left") +
  xlab("No. of Policies to Encourage Physical Activity at the State-Level")+
  ylab("No. of States")+
  ggtitle("No. of Policies to Encourage Physical Activity at the State-Level, 2018") +
  scale_x_continuous(breaks = seq(0 , 7 , by=1)) +
  ylim(c(0,35)) +
  stat_bin(binwidth=1,center=1,geom="text", aes(label=..count..), vjust=-1.5, size=3.5)+
  theme_minimal()

plot(hist2)
```

Most states have exacty 1 Policy that Encourages Physical Activity. Around one-third of the states have no policy at all.

``` {r, fig.height=6, fig.width=6}

# Television
questions_tel_yesnoq <- c("State-National Alignment on Restricting Media Age >2 yrs","State-National Alignment on Prohibiting Media Age <=2 yrs")
pe_2018_tel_yesnoq <- pe_2018_states %>%
  filter(Class=="Television Viewing" & QuestionRenamed %in% questions_tel_yesnoq) %>%
  group_by(LocationAbbr) %>%
  summarise(n_policies = sum(Data_Value=="Yes"))

hist3 <- ggplot(pe_2018_tel_yesnoq, aes(x= n_policies)) +
  geom_histogram(binwidth=1,center=1, colour = "black", fill="#f1beb6", closed="left") +
  xlab("No. of Policies to Restrict TV Viewing at the State-Level")+
  ylab("No. of States")+
  ggtitle("No. of Policies to Restrict TV Viewing at the State-Level, 2018") +
  scale_x_continuous(breaks = seq(0 , 7 , by=1)) +
  stat_bin(binwidth=1,center=1,geom="text", aes(label=..count..), vjust=-1.5, size=3.5)+
  theme_minimal()

plot(hist3)
```

Very few states have any policy regarding Restricting Television Viewing (7). All the other states have no such policies.

``` {r, fig.height=10, fig.width=10}

# All policies for yes or no per state
questions_yesnoq <- c(questions_nutrition_yesnoq,questions_pa_yesnoq,questions_tel_yesnoq)

pe_2018_yesnoq <- pe_2018_states %>%
  filter(QuestionRenamed %in% questions_yesnoq) %>%
  group_by(LocationAbbr,Class) %>%
  summarise(n_policies = sum(Data_Value=="Yes"))

pe_2018_yesnoq$state = pe_2018_yesnoq$LocationAbbr

pe_2018_yesnoq_wider <- pe_2018_yesnoq %>% pivot_wider(names_from = Class, values_from = n_policies)

pe_2018_yesnoq_wider$Total <- pe_2018_yesnoq_wider$`Fruits and Vegetables` + pe_2018_yesnoq_wider$`Physical Activity`+
  pe_2018_yesnoq_wider$`Television Viewing`

pe_2018_yesnoq_wider_pivot <- pivot_longer(pe_2018_yesnoq_wider, `Fruits and Vegetables`:`Total`, names_to = "Class", values_to = "n_policies")

pe_2018_yesnoq_wider_pivot[is.na(pe_2018_yesnoq_wider_pivot)] <- 0


sb1 <- ggplot(pe_2018_yesnoq_wider_pivot, aes(state=LocationAbbr, fill=n_policies)) +
        geom_statebins() +
        coord_equal() +
        viridis::scale_fill_viridis(
        name = "Number of policies ", limits=c(0,6), breaks=0:6, option = "magma", direction = -1
        ) +
        facet_wrap(~Class, ncol=2) +
        labs(title="Number of policies for good Nutrition, Physicial Activity and Television Viewing ") +
        theme(plot.title=element_text(size=16, hjust=0)) +
        theme(plot.margin = margin(1,1,1,1)) +
        theme(axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank())

plot(sb1)

```

From the histograms, we observed the distribution of number of policies at the state-level.
From this geo-plot, we can see exactly which state has how many policies supporting each of the Classes - Nutrition, Physical Activity and Television Viewing.

From the right-bottom graph, which sums up total number of policies from all three Classes per state, we observe that it is mostly Southern states like New Mexico, Tennessee, Mississippi, North Carolina etc. with the highest number of policies.
There could be multiple possible explanations for these higher numbers. It is possible that these States have the highest obesity rates in the US, and therefore the US government is more active in establishing policies to promote heath. An alternative explanation could be that 
these states have shown promise in reducing obesity rates as a result of the implementation of these policies, and therefore the government is more active in this regard. 

```{r, fig.height=10, fig.width=10}

pe_2018_breastfeeding_questions <- pe_2018_states %>%
  filter(Class=="Breastfeeding")%>%
  select(QuestionRenamed) %>% distinct()

pe_2018_states$Data_Value <- as.numeric(pe_2018_states$Data_Value)

pe_2018_breastfeeding_questions_data <- pe_2018_states %>%
  filter(QuestionRenamed %in% pe_2018_breastfeeding_questions$QuestionRenamed)%>%
  group_by(LocationDesc) %>%
  summarise(avg_brf = mean(Data_Value))

pe_2018_breastfeeding_questions_data[is.na(pe_2018_breastfeeding_questions_data)] <- 0

df_pe <- pe_2018_breastfeeding_questions_data %>% as.data.frame() %>% 
    transmute(region = tolower(LocationDesc), value = avg_brf)

state_choropleth(df_pe,
                 title = "State Breastfeeding Support, 2018",
                 legend = "Average Breastfeeding Support")
```

We also had data regarding Breastfeeding Support in terms of the "Maternity Practices in Infant Nutrition and Care (mPINC)" score as well as the Percent of live births occurring at facilities designated as "baby friendly" by the Baby Friendly Hospital Initiative (BFHI). Since the mPINC score is also out of 100, we took an average of these scores per state to plot this graph.

We do not observe a particular region in the US to be doing significantly better than others in providing Policy Support For BreastFeeding.

From this dataset, our broad conclusions are:

1. At the State-level, among different Classes of Policy Support, the most number of policies implemented are found to be encouraging Good Nutrition Intake, followed by Physical Activity and lastly Television Viewing, which is hardly restricted in the US.

2. Breastfeeding Policy Support Data across states does not seem to follow a discernable pattern in terms of more or less support in a particular US region.

Broad conclusion from analysis of both datasets:

There seems to be a somewhat positive correlation between Obesity Rates and Number of Policies supporting good health, implemented by the US government. This is mostly observed in the southern states. 

References

[1]  Link: [CDC - Proven Strategies for health] https://www.cdc.gov/nccdphp/dnpao/proven-strategies.html "CDC - Proven Strategies for health"
