==> 06-interactive.Rmd <==
# Interactive component

In this component, we do a time-series plot of Obesity Status, Physical Activity Status, and Nutrition Status at the State-level instead of the National level (done in section 2).

The first plot shows a map of the US with the colour of each state [1] indicating the level of obesity in that state, in 2019. To observe the trends for a particular state for the past decade, we can click on a point within the state area on the map, and then a time series plot appears below. The plot consists of trend lines for Obesity Status, Physical Activity Status, and Nutrition Status. Physical Activity Status is shown using the "% No Physical Activity" which translates percentage of population not engaging in any physical activity in a week. Similarly, Nutrition Status is indicated using "% No Fruits/Vegetables" which is basically the percentage of population not consuming any fruits or vegetables on a daily basis

We can observe that there are some states with fairly lower rates of increase in obesity levels over the years. 

<script src="https://d3js.org/d3.v6.js"></script>
<div id="plot1"></div>
<div id="plot2"></div>

<script src="scripts/maps.js"></script


References 

[1] Link : [Basic US State Map - D3](http://bl.ocks.org/michellechandra/0b2ce4923dc9b5809922 "Basic US State Map - D3")
