==> 07-conclusion.Rmd <==
# Conclusion

### Limitations

In the "Dataset 1 Obesity-Behavioral" dataset, there was a column named "Sample Size" which had a different value for every Stratification provided in the dataset. This meant that since it was not a single sample from which all the numbers in the data were derived. Therefore we could not perform any analysis to see the proportions of population belonging to each of the Stratifications within the StratificationCategories. This analysis would have been useful to understand which strata of the society are contributing the most to higher obesity levels over the years. For example, are the overlapping groups in lower income and lower educational levels groups contributing the most?

Due to the sparsity of "Dataset 2 Obesity-Policy-Support" across the years, in the past decade, we were unable to analyse time-series trends for the policies supporting the different classes - BreastFeeding, Nutrition Intake, Physical Activity and Television Viewing. We were unable to comment if there was association of policies being implemented versus obesity trends over the years i.e. did more policies help to reduce obesity rates in that state, over the years?  

### Future Scope

We could aim to root cause underlying factors responsible for the visible gaps observed in Obesity Rates among sub-cateogires of StratificationCategories. For example, Asians were observed to have a lower obesity rate although their "Lack of Fruit and Vegetable Consumption" and "Lack of Physical Activity" levels were comparable to other "Race/Ethnicity" groups which had much higher obesity rates.
Factors like genetics, underlying health conditions of the sub-group etc. could possible provide explanations for such observations. We could join our data with datasets providing valuable information about these. 

We could research associations between COVID hospitalizations, deaths, intensity of COVID symptoms etc. with this obesity data using additional COVID-related datasets.

In Section 6, we noted that states did show some variations in levels of increase in Obesity Status, Physical Activity Status, and Nutrition Status over the years. It would be interesting to broadly categorise states into 2 categories - states with steep increases in obesity rates over the years, almost constant to low levels of increase in obesity rates. It would be then worth taking a deeper dive into these states and exploring the reasons behind these differences.
